+++ src/components/splashScreen.vue
@@ -11,17 +11,29 @@
           <img
             v-if="darkTheme"
             class="logo"
-            src="../../public/img/weektodo-isologo-white.svg"
-            alt="WeekToDo Logo"
+            src="../../public/img/Weeklyst-isologo-white.svg"
+            alt="Weeklyst Logo"
             style="display: inline"
           />
           <img
             v-else
             class="logo"
-            src="../../public/img/weektodo-isologo-color.svg"
-            alt="WeekToDo Logo"
+            src="../../public/img/Weeklyst-isologo-color.svg"
+            alt="Weeklyst Logo"
             style="display: inline"
           />
+        </div>
+        <div v-show="sponsor" class="d-flex justify-content-center" style="height: 50px">
+          <a v-if="sponsor" :href="sponsor.url" class="d-flex sponsor-container align-items-center" target="_blank">
+            <img :src="sponsor.img" class="sponsor-img" alt="Weeklyst Sponsor" />
+            <div class="my-2 mx-2">
+              <div class="fw-bolder d-inline" style="text-decoration: unset !important">{{ sponsor.name }}</div>
+              <div class="opacity-50 mx-2 d-inline">{{ sponsor.message }}</div>
+            </div>
+          </a>
+        </div>
+        <div style="height: 25px; width: 200px" class="d-flex justify-content-center">
+          <div v-show="sponsor" class="opacity-25" style="font-size: 0.7rem">{{ $t("ui.sponsoredBy") }}</div>
         </div>
       </div>
     </div>
@@ -41,7 +53,7 @@ export default {
   mounted() {
     const axios = require("axios").default;
     axios
-      .get("https://weektodo.me/api/sponsors")
+      .get("https://weeklyst.com/api/sponsors")
       .then((response) => this.renderSponsor(response))
       .catch((error) => console.log(error.message));
   },
@@ -53,9 +65,9 @@ export default {
       var sponsors = [];
 
       sponsors.push({
-        name: "WeekToDo",
+        name: "Weeklyst",
         message: this.$t("donate.splashMessage"),
-        url: "https://weektodo.me/support-us",
+        url: "https://weeklyst.com/support-us",
         img: "/icons/ko-fi.png",
       });